% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.collection}
\alias{as.collection}
\title{As collection}
\usage{
as.collection(x, endpoint = as.endpoint("PlanetaryComputer"))
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Format to STAC collection name for some commonly used collections and STAC endpoints
}
