% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.endpoint}
\alias{as.endpoint}
\title{As endpoint}
\usage{
as.endpoint(name)
}
\arguments{
\item{name}{character}
}
\value{
a character
}
\description{
Some endpoints of STAC APIs because I'm too lazy to remember/search
}
\details{
Defaults to Planetary Computer
}
