% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.SpatRaster}
\alias{as.SpatRaster}
\title{Coerce to SpatRaster}
\usage{
as.SpatRaster(x)
}
\arguments{
\item{x}{data cube}
}
\value{
SpatRaster
}
\description{
Writes a data cube as a temporal file and reads that as SpatRaster object
}
