% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets2rast.R
\name{assets2rast}
\alias{assets2rast}
\title{STAC assets to SpatRaster}
\usage{
assets2rast(feature, assets, as_list = FALSE)
}
\arguments{
\item{feature}{a STAC feature: list element of STACItemCollection$features}

\item{assets}{character vector of assets names to extract}

\item{as_list}{boolean defining whether assets should be read as list of SpatRaster objects or as a single multilayer SpatRaster object}
}
\value{
a (list of) SpatRaster object(s)
}
\description{
Read selected assets from a STAC feature as a SpatRaster object
}
\details{
Assets to read must be of same extent and dimensions if read into mulitlayer SpatRaster. Set 'as_list=TRUE' to deal with assets with different extents or dimensions.
Assets must be readable, which may require prior authentication.
}
\seealso{
[\code{\link{assets2vrt}}]
}
