% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets2rast.R
\name{assets2vrt}
\alias{assets2vrt}
\title{STAC assets to virtual raster}
\usage{
assets2vrt(featureCollection, assets, as_list = FALSE)
}
\arguments{
\item{assets}{character vector of assets names to extract}

\item{as_list}{boolean defining whether assets should be read as list of SpatRaster objects or as a single multilayer SpatRaster object}

\item{feature}{a STAC feature: list element of STACItemCollection$features}
}
\value{
a (list of) SpatRaster object(s)
}
\description{
Read selected assets from a STAC featureCollection as a virtual SpatRaster object
}
\details{
Can be used when the different features in the featureCollection are tiles that must be combined.
}
\seealso{
[\code{\link{assets2rast}}]
}
