% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI_composite.R
\name{BA_VI_ts}
\alias{BA_VI_ts}
\title{Before/After Vegetation Index time series}
\usage{
BA_VI_ts(
  spatRef,
  index,
  collection,
  endpoint,
  intervention_year,
  intervention_lag = 0,
  nYears,
  nYears_before,
  nYears_after,
  months = 1:12,
  maxCloud = 60
)
}
\arguments{
\item{spatRef}{SpatRaster defining spatial reference for output layers}

\item{index}{Character. Vegetation Index to be calculated}

\item{collection}{description}

\item{endpoint}{STAC endpoint}

\item{intervention_year}{Numeric. Year of intervention}

\item{intervention_lag}{Numeric. Lag (in years) of response to intervention}

\item{nYears}{Numeric. Length of the before and after time series in years}

\item{nYears_before}{Numeric. Length of the before time series in years. Ignored if argument nYears is provided}

\item{nYears_after}{Numeric. Length of the after time series in years. Ignored if argument nYears is provided}

\item{months}{Numeric vector. Month(s) for which the yearly composites should be created}

\item{maxCloud}{Numeric. Maximum cloud cover of images used in compositing}

\item{name}{description}
}
\value{
List of SpatRaster objects. The two list elements correspond to the before and after period, respectively
}
\description{
Create yearly composite time series for the before and after period for an intervention
}
