% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_eval.R
\name{BACI_contrast}
\alias{BACI_contrast}
\title{Before-After-Control-Impact contrast}
\usage{
BACI_contrast(data, SpatRef = NULL)
}
\arguments{
\item{data}{data.table}

\item{SpatRef}{SpatRaster}
}
\value{
data.table
}
\description{
Calculate the Control-Impact contrast and p-value
}
\details{
After Meroni et al. (2017) and del Río-Mena et al. (2021)
}
\references{
Meroni, M., Schucknecht, A., Fasbender, D., Rembold, F., Fava, F., Mauclaire, M., Goffner, D., Di Lucchio, L.M., Leonardi, U., 2017. Remote sensing monitoring of land restoration interventions in semi-arid environments with a before–after control-impact statistical design. International Journal of Applied Earth Observation and Geoinformation 59, 42–52.

del Río-Mena, T., Willemen, L., Vrieling, A., Snoeys, A., Nelson, A., 2021. Long-term assessment of ecosystem services at ecological restoration sites using Landsat time series. PLOS ONE 16, e0243020.
}
\seealso{
[\code{\link{CI_contrast}}]
}
