% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baviaanskloof_roads-data.R
\docType{data}
\name{baviaanskloof_roads}
\alias{baviaanskloof_roads}
\title{Baviaanskloof roads}
\format{
A SpatVector object
}
\usage{
data(baviaanskloof_roads)
}
\description{
Subset of the [OpenStreetMap](https://www.openstreetmap.org/copyright) roads dataset for the area around Baviaankloof. Downloaded from download.geofabrik.de.
}
\examples{
data(baviaanskloof_roads)
baviaanskloof <- unwrap(baviaanskloof_roads)
plot(baviaanskloof_roads)

}
\keyword{datasets}
