% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_mask.R
\name{bitmask_values}
\alias{bitmask_values}
\title{Landsat L2 QA bitmap values}
\usage{
bitmask_values(
  nBit = 8,
  fill = TRUE,
  dilCloud = TRUE,
  cirrus = TRUE,
  cloud = TRUE,
  cloudShadow = TRUE,
  snow = FALSE,
  water = FALSE
)
}
\arguments{
\item{nBit}{integer. Number of bits from bitmask used, defaults to 8}

\item{fill}{logical. Mask fill bit}

\item{dilCloud}{logical. Mask dilated cloud bit}

\item{cirrus}{logical. Mask cirrus bit}

\item{cloud}{logical. Mask cloud bit}

\item{cloudShadow}{logical. Mask cloud shadow bit}

\item{snow}{logical. Mask snow bit}

\item{water}{logical. Mask water bit}
}
\value{
integer vector
}
\description{
Define the values of values to be masked from Landsat L2 products based on specified QA bits
}
