% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_eval.R
\name{calc_contrast_p}
\alias{calc_contrast_p}
\title{Calculate contrast and p value}
\usage{
calc_contrast_p(
  data,
  colname.subclass = "subclass",
  colname.treatment = "treatment"
)
}
\arguments{
\item{data}{data.table or data.frame. See details.}

\item{colname.subclass}{character. Name of column indicating subclass}

\item{colname.treatment}{character. Name of column indicating treatment}
}
\description{
Calculate the (BA)CI contrast and p-value for matched control-impact units
}
\details{
Parameter \code{data} must be a data.table (or data.frame) with columns indicating the subclass and treatment, the remaining columns will be assumed to be the effects for which contrast and p-value must be calculated
}
