% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_metrics.R
\name{calc_ts_metrics}
\alias{calc_ts_metrics}
\title{Calculate time series metrics}
\usage{
calc_ts_metrics(x, ...)
}
\arguments{
\item{x}{data cube or spatRaster}

\item{...}{additional arguments passed to \code{calc_ts_metrics.cube} or \code{calc_ts_metrics.rast}}
}
\description{
Generic function for \code{calc_ts_metrics.cube} and \code{calc_ts_metrics.spatRaster}
}
\details{
Currently, there are some differences in the inputs/outputs of the functions for data cube and spatRaster, this has to be harmonized.
}
