% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_metrics.R
\name{calc_ts_metrics.SpatRaster}
\alias{calc_ts_metrics.SpatRaster}
\title{Calculate time series metrics from a SpatRaster}
\usage{
\method{calc_ts_metrics}{SpatRaster}(rastTimeSeries, average = 1, trend = 1, immediate = 0)
}
\arguments{
\item{rastTimeSeries}{SpatRaster. Each layer represents a time step, provided in "time" field}

\item{average}{numeric. If not 0, calculate the time series average}

\item{trend}{numeric. If not 0, calculate the time series trend}

\item{immediate}{numeric. If not 0, calculate the time series component required to calculate the immediate effect. Provide negative value for the "before", positive value for the "after" time series.}
}
\value{
SpatRaster with selected parameters as layers
}
\description{
Calculate average and trend from a SpatRaster times series
}
