% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VI.R
\name{calc_VI.cube}
\alias{calc_VI.cube}
\title{Calculate Vegatation Index}
\usage{
calc_VI.cube(cube, VI, scale = 1e-05)
}
\arguments{
\item{cube}{data cube with spectral bands}

\item{VI}{vegetation index}

\item{scale}{scale of source surface reflectance data}
}
\value{
proxy data cube object
}
\description{
Calculates Vegetation Index for a data cube of EO (surface reflectance) images
}
\details{
Implemented vegetation indices:
\itemize{
  \item NDVI = (nir-red)/(nir+red)
  \item NDMI = (nir-swir1)/(nir+swir1)
  \item NBR = (nir-swir2)/(nir+swir2)
  \item NBR2 = (swir1-swir2)/(swir1+swir2)
  \item NDSI = (green-swir1)/(green+swir1)
  \item EVI = 2.5*(nir-red)/(nir+6*red-7.5*blue+1)
  \item SAVI = 1.5*(nir-red)/(nir+red+0.5)
  \item MSAVI = (2*nir+1-sqrt((2*nir+1)^2-8*(nir-red)))/2
}

Scale factor has to be applied when reflectance saved as integer instead of range [0-1]. To be checked for different endpoints
}
