% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_eval.R
\name{CI_contrast}
\alias{CI_contrast}
\title{Control-Impact contrast}
\usage{
CI_contrast(data, SpatRef = NULL)
}
\arguments{
\item{data}{data.table}

\item{SpatRef}{SpatRaster}
}
\value{
data.table
}
\description{
Calculate the Control-Impact contrast and p-value
}
\seealso{
[\code{\link{BACI_contrast}}]
}
