% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control_candidates.R
\name{create_control_candidates}
\alias{create_control_candidates}
\title{Create candidate control raster}
\usage{
create_control_candidates(
  impact,
  resolution,
  crs,
  control_from_buffer = NULL,
  control_from_include = NULL,
  control_exclude = NULL,
  exclude_impact_buffer = NULL,
  round_coords = FALSE
)
}
\arguments{
\item{impact}{spatVector. Vector representation (points or polygons) of the impact sites}

\item{resolution}{numeric. Spatial resolution of the output spatRaster}

\item{crs}{Coordinate Reference System in PROJ.4, WKT or authority:code notation. Defaults to the UTM zone of center coordinates of \code{impact}}

\item{control_from_buffer}{numeric vector of 1, 2, or 4 elements. Indicates by how much (in units of crs) the spatial extent of x/y should be enlarged on each side.}

\item{control_from_include}{spatVector. Indicates area to include in candidate control selection (optional)}

\item{control_exclude}{spatVector. Indicates area to exclude from candidate control selection}

\item{round_coords}{logical or integer. Should the coordinates of the reference SpatRaster be rounded.
If TRUE, coordinates are rounded to the nearest integer; if a positive numeric rounded to the corresponding decimal; 
if a negative integer rounded to the corresponding power of 10 (e.g., \code{round_corrs=2} round to the nearest 100).}
}
\value{
a spatRaster in which impact pixels have value 1, candidate control pixels value 0, and all other pixels value NA
}
\description{
Creates a spatRaster of impact and candidate control pixels 
 
Impact pixels are defined by providing a spatVector object. Candidate control pixels can be defined from a spatVector file, or using a buffer around the bounding box of the impact sites.
}
