% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube2SpatRaster.R
\name{cube2SpatRaster}
\alias{cube2SpatRaster}
\title{Data cube to SpatRaster}
\usage{
cube2SpatRaster(cube)
}
\arguments{
\item{cube}{data cube object}
}
\value{
a SpatRaster object
}
\description{
Convert a proxy data cube object to
}
\details{
Probably only works with a 3-dimensional data cube, to be checked
}
