% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_cube.R
\name{cube_composite.median}
\alias{cube_composite.median}
\title{Create a median value composite image}
\usage{
cube_composite.median(cube)
}
\arguments{
\item{cube}{source data cube}
}
\description{
Create a median value composite image from a data cube
}
