% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_composite.R
\name{cube_composite.medoid}
\alias{cube_composite.medoid}
\title{Create medoid composite image}
\usage{
cube_composite.medoid(cube)
}
\arguments{
\item{cube}{source data cube}
}
\value{
output data cube
}
\description{
Create a pseudo-medoid composite image from a data cube
}
\details{
This compositing method uses aan approximating of the medoid, selecting the observation with the smallest euclidean distance to the band-wise median position.
}
