% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_composite.R
\name{cube_composite}
\alias{cube_composite}
\title{Composite cube}
\usage{
cube_composite(cube, method = "median", ...)
}
\arguments{
\item{cube}{a multitemporal data cube}

\item{method}{compositing method: "median" or "medoid"}

\item{...}{optional arguments for specific methods}
}
\value{
data cube
}
\description{
Create a pixel-based composite image from a multitemporal data cube
}
\details{
See cube_composite.median, cube_composite.medoid
}
