% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeVI.R
\name{cubeVI}
\alias{cubeVI}
\title{Calculate Vegatation Index}
\usage{
cubeVI(cube, VI, endpoint, collection, scale = 1e-05)
}
\arguments{
\item{cube}{data cube with spectral bands}

\item{VI}{vegetation index}

\item{endpoint}{STAC endpoint}

\item{collection}{STAC collection}

\item{scale}{of source surface reflectance data}
}
\value{
proxy data cube object
}
\description{
Calculates Vegetation Index for a data cube of EO (surface reflectance) images
}
\details{
Implemented vegetation indices:
*NDVI = (nir-red)/(nir+red)
*EVI
*SAVI
*MSAVI
*NDMI
*NBR
*NBR2
*NDSI

Scale for VIs with constants has to be checked for different endpoints
}
\seealso{
[reqBands()]
}
