% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{dem}
\alias{dem}
\title{Extract Digital Elevation Model}
\usage{
dem(x, ...)
}
\arguments{
\item{x}{a SpatRaster, spatVector, or SpatExtent object, defining the area for which the DEM should be extracted}

\item{dx}{numeric. output resolution in x dimension}

\item{dy}{numeric. output resolution in y dimension}

\item{srs}{character. output spatial reference system}

\item{dem_source}{Source of the Digital Elevation Model, see Details}

\item{v}{character vector. Terrain parameters to be calculated}

\item{neighbors}{integer. Either 8 (queen case) or 4 (rook case), indicating which neighbors to use to compute slope or aspect.}

\item{unit}{character. "degrees" or "radians" for the output of "slope" and "aspect"}

\item{transformAspect}{logical. Should aspect be transformed to "northness" and "eastness"}
}
\value{
a data cube or SpatRaster object (see Details)
}
\description{
Read Digital Elevation Model and calculate derivatives.
}
\details{
This function returns a data cube object when only the digital terrain model is required (i.e. \code{v="elevation"}), and a SpatRaster when the terrain derivatived are required.
The reason for this is that calculating terrain parameters using the terra::terrain function outperforms the same calculation on data cube, with regard to time.

\code{dem_source} must be a list with STAC specifications (named list with list elements "endpoint", "collection", "assets" and (optionally) "authOpt".
If no DEM source it provided, this argument defaults to \code{list(endpoint="https://planetarycomputer.microsoft.com/api/stac/v1", collection="nasadem", assets="elevation", authOpt=list())}

Available DEMs for selected STAC endpoints:
\itemize{
\item endpoint="https://planetarycomputer.microsoft.com/api/stac/v1"
}\tabular{llll}{
   name \tab collection= \tab assets= \tab auth. required \cr
   NASA DEM \tab "nasadem" \tab "elevation" \tab no \cr
   ALOS World 3D-30m \tab "alos-dem" \tab "data" \tab no \cr
   Copernicus DEM GLO-30 \tab "cop-dem-glo-30" \tab "data" \tab yes \cr
   Copernicus DEM GLO-90 \tab "cop-dem-glo-90" \tab "data" \tab yes \cr
}
}
