% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{dem2cube}
\alias{dem2cube}
\title{Read DEM as data cube}
\usage{
dem2cube(
  xmin,
  xmax,
  ymin,
  ymax,
  dx,
  dy,
  srs,
  endpoint,
  collection,
  assets,
  authOpt = list(),
  ...
)
}
\arguments{
\item{xmin}{numeric. Minimum x-coordinate of output DEM}

\item{xmax}{numeric. Maximum x-coordinate of output DEM}

\item{ymin}{numeric. Minimum y-coordinate of output DEM}

\item{ymax}{numeric. Maximum y-coordinate of output DEM}

\item{dx}{numeric. Pixels size of output DEM in x-dimension}

\item{dy}{numeric. Pixels size of output DEM in y-dimension}

\item{srs}{character. Description of spatial reference system}

\item{endpoint}{character. STAC endpoint}

\item{collection}{character. STAC collection}

\item{assets}{character. STAC asset name}

\item{authOpt}{STAC authentication options (endpoint-dependent)}

\item{...}{additional arguments (not used)}
}
\value{
a gdalcubes data cube
}
\description{
Read a DEM from a STAC collection or from file as a gdalcubes data cube
}
