% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToVector.R
\name{distanceToVector}
\alias{distanceToVector}
\title{Geographic distance to feature of interest}
\usage{
distanceToVector(x, y, ...)
}
\arguments{
\item{x}{SpatRaster or Spatvector for which distances must calculated}

\item{y}{a terra::SpatVector or sf::sf vector layer (points, lines or polygons) to which distances must be calculated}

\item{...}{Additional arguments to terra::distance}
}
\value{
SpatRaster or SpatVector
}
\description{
Calculates (minimum) geographic distance in meters from the cells of a reference SpatRaster or the feature in reference SpatVector to the features in the target SpatVector
}
\details{
Basically rast::distance but with reprojection CRS if needed, and in case of SpatVector-SpatVector returning the minimum distance rather than a matrix.
}
