% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_format.R
\name{eo_assets}
\alias{eo_assets}
\title{EO assets}
\usage{
eo_assets(feature, collection, meta = TRUE)
}
\arguments{
\item{feature}{description}

\item{collection}{description}

\item{meta}{description}
}
\value{
character vector
}
\description{
Default assets for L/S2
}
\details{
To be updated for correct naming under PC/LPDAAC/..., For now only PC Landsat and Sentinel-2 implemented
}
