% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_cube.R
\name{eo_cube.stac}
\alias{eo_cube.stac}
\title{EO cube from STAC}
\usage{
eo_cube.stac(
  x,
  endpoint,
  collection,
  t0,
  t1,
  dt = "P1D",
  asset_names = NULL,
  maxCloud = NULL,
  maskOptions = list(NULL),
  stacOptions = list(limit = 1000, authOpt = list()),
  aggregation = "first",
  resampling = "bilinear"
)
}
\arguments{
\item{x}{spatRaster. Spatial reference}

\item{endpoint}{STAC endpoint}

\item{collection}{STAC collection}

\item{t0}{t0}

\item{t1}{t1}

\item{dt}{dt}

\item{asset_names}{description}

\item{maxCloud}{description}

\item{maskOptions}{description}

\item{stacOptions}{Optional arguments to be passed to eo_stac_search, e.g. authentication options}

\item{resampling}{description}

\item{aggregations}{description}
}
\description{
Create a gdalcubes data cube for EO images retrieved through STAC
}
\details{
Warning: masking only works now when asset_names are explicitely defined
}
