% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_cube.R
\name{eo_cube_yearly.stac}
\alias{eo_cube_yearly.stac}
\title{Yearly EO cube from STAC}
\usage{
eo_cube_yearly.stac(
  x,
  endpoint,
  collection,
  stacOptions = list(),
  years,
  months,
  asset_names = NULL,
  joint_assets = TRUE,
  maxCloud = NULL,
  maskOptions = list(NULL),
  aggregation = "median",
  resampling = "bilinear",
  harm_bandnames = TRUE
)
}
\arguments{
\item{x}{spatRaster. Defines spatial reference of the output data cube}

\item{endpoint}{character. STAC endpoint}

\item{collection}{character. STAC collection}

\item{stacOptions}{list. Optional arguments to be passed to eo_stac_search, e.g. authentication options}

\item{years}{description}

\item{months}{description}

\item{asset_names}{character vector of asset (band) names to be used, or NULL for all asset names with "eo:bands" attributes}

\item{joint_assets}{logical indicating if only the assets (bands) present in all features should be used, if argument asset_names is NULL}

\item{maxCloud}{numeric. Maximum cloud cover [0-100] of images to be included in data cube, only relevant for collections with "eo:cloud_cover" attributes}

\item{maskOptions}{list specifying layer used to create mask and masking options (see eo_mask.cube)}

\item{aggregation}{see gdalcubes::cube_view}

\item{resampling}{see gdalcubes::cube_view}

\item{harm_bandnames}{logical indicating whether endpoint/collection-specific band names should be changed to generic band names}
}
\value{
proxy data cube object
}
\description{
Create a data cube with yearly time intervals for EO images retrieved from a STAC catalog.
}
\details{
This function bypasses limits on the length of the queried STAC itemCollection by posting subsequent queries for each year
}
