% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_harm_bandnames.R
\name{eo_harm_bandnames.cube}
\alias{eo_harm_bandnames.cube}
\title{Harmonize STAC endpoint/collection-specific band names of data cubes}
\usage{
eo_harm_bandnames.cube(cube, endpoint, collection)
}
\arguments{
\item{cube}{a data cube}

\item{endpoint}{character. STAC endpoint}

\item{collection}{character. STAC collection}
}
\description{
Replaces the STAC endpoint/collection-specific band names of a data cube by generic band names (i.e., "blue", "green", "red", "nir", ...).
This simplifies the calculation of Vegetation Indices,
}
