% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_harm_bandnames.R
\name{eo_harm_bandnames}
\alias{eo_harm_bandnames}
\title{Harmonize STAC endpoint/collection specific band names}
\usage{
eo_harm_bandnames(x, ...)
}
\arguments{
\item{x}{data cube}

\item{...}{additional arguments passed to \code{eo_harm_bandnames.cube}}
}
\description{
Harmonize STAC endpoint/collection specific band names
}
