% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmask_values.R
\name{eo_image_mask}
\alias{eo_image_mask}
\title{EO mask for gdalcubes}
\usage{
eo_image_mask(maskOptions = list(maskLyr = NULL))
}
\arguments{
\item{maskOptions}{named list. Masking options. See details.}
}
\description{
Create mask for gdalcube from Landsat QA layer or Sentinel-2 SCL
}
\details{
Clouds+cloud shadows always masked. Snow and water defined in argument maskOptions (default=FALSE, for both snow and water)
Possible list names: maskLyr, maskSnow, maskWater
}
