% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_harm_bandnames.R
\name{eo_mapBand}
\alias{eo_mapBand}
\title{Map band}
\usage{
eo_mapBand(bandnames, endpoint, collection, invert = FALSE)
}
\arguments{
\item{bandnames}{character vector of bandnames}

\item{endpoint}{character. STAC endpoint}

\item{collection}{character. STAC collection}

\item{invert}{logical. Set to TRUE to obtain endpoint-specific bandnames from generic names}
}
\value{
character vector
}
\description{
Switch between generic and endpoint/collection-specific band name
}
\details{
Limited number of endpoints/collections implemented, more to be added
}
