% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_mask.R
\name{eo_mask.cube}
\alias{eo_mask.cube}
\title{EO mask for data cube}
\usage{
eo_mask.cube(maskLyr, maskSnow = FALSE, maskWater = FALSE)
}
\arguments{
\item{maskLyr}{character. Name of mask layer.}

\item{maskSnow}{logical. Defines whether snow should be masked.}

\item{maskWater}{logical. Defines whether water should be masked.}

\item{maskOptions}{named list. Masking options. See details.}
}
\value{
an object of class "image_mask"
}
\description{
Create mask from Landsat QA layer or Sentinel-2 SCL that can be used in gdalcubes::raster_cube
}
\details{
The metadata layer from which the mask is derived should be specified in \code{maskLyr}. For now, mask derived from "qa_pixel" (for Landsat) and "SCL" (for Sentinel-2) are implemented.
Clouds, cloud shadow and pixels adjacent to clouds are always masked (when these fields are specified in the masking layer). 
Snow and water can be masked by setting the respective parameters, by default these are not masked
}
\examples{
# mask from Landsat "qa_pixel" layer
eo_mask.cube("qa_pixel")
# mask from Sentinel-2 "SCL" layer
eo_mask.cube("SCL", maskSnow=TRUE, maskWater=TRUE)

}
