% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_mask.R
\name{eo_maskLyrName.stac}
\alias{eo_maskLyrName.stac}
\title{Retrieve mask layer name}
\usage{
eo_maskLyrName.stac(endpoint, collection)
}
\arguments{
\item{endpoint}{STAC endpoint}

\item{collection}{STAC collection}
}
\value{
character or NULL
}
\description{
This function returns the metadata layer name from which a mask can be derived for a collection at a STAC endpoint
}
\details{
Currently only implemented for collections "landsat-c2-l2" and "sentinel-2-l2a" at Planetary Computer. Returns NULL for all other values. Additional endpoints/collections will be added.
}
