% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_stac_composite.R
\name{eo_stac_composite}
\alias{eo_stac_composite}
\title{EO STAC composite}
\usage{
eo_stac_composite(
  spatRef,
  t0,
  t1,
  endpoint,
  collection,
  limit = 1000,
  maxCloud = 100,
  assets = eo_assets(collection = collection),
  maskSnow = FALSE,
  maskWater = FALSE,
  aggregation = "median",
  resampling = "bilinear"
)
}
\arguments{
\item{spatRef}{SpatRaster object from which spatial reference is extracted}

\item{t0}{composite start time, character or Date formatted as "yyyy-mm-dd",  or numeric formatted as yyyymmdd}

\item{t1}{composite end time, character or Date formatted as "yyyy-mm-dd", or numeric formatted as yyyymmdd}

\item{endpoint}{STAC endpoint}

\item{collection}{STAC collection}

\item{limit}{STAC item limit}

\item{assets}{STAC assets to be selected}

\item{maskSnow}{should snow be masked if mask layer is available}

\item{maskWater}{should water be masked if mask layer is available}

\item{aggregation}{aggregation method, see \code{gdalcubes::cubes_view}}

\item{resampling}{resampling method, see \code{gdalcubes::cubes_view}}

\item{maxcloud}{maximum cloud cover}
}
\value{
proxy data cube object
}
\description{
Create EO image composite from STAC collection
}
\details{
Automatically applies masking if masking layer is provided
}
