% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_stac_search.R
\name{eo_stac_search}
\alias{eo_stac_search}
\title{Search STAC EO data}
\usage{
eo_stac_search(endpoint, ...)
}
\arguments{
\item{endpoint}{see [rstac::search()]}

\item{...}{see [eo_stac_search.PlanetaryComputer()]}
}
\value{
STACItemCollection
}
\description{
Easier searching for EO data in STAC collections
}
\details{
For now only implemented for Planetary Computer
}
