% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_stac_composite.R
\name{eo_stac_yearly_composites}
\alias{eo_stac_yearly_composites}
\title{Yearly STAC composite}
\usage{
eo_stac_yearly_composites(
  spatRef,
  years,
  months,
  endpoint,
  collection,
  limit = 1000,
  maxCloud = 100,
  assets = eo_assets(collection = collection),
  maskSnow = FALSE,
  maskWater = FALSE,
  aggregation = "median",
  resampling = "bilinear",
  authOpt = list()
)
}
\arguments{
\item{spatRef}{SpatRaster object from which spatial reference is extracted}

\item{years}{years for which a composite is created}

\item{months}{months that are used for creating composite for each year}

\item{endpoint}{STAC endpoint}

\item{collection}{STAC collection}

\item{limit}{STAC item limit}

\item{assets}{STAC assets to be selected}

\item{maskSnow}{should snow be masked if mask layer is available}

\item{maskWater}{should water be masked if mask layer is available}

\item{aggregation}{aggregation method, see \code{gdalcubes::cubes_view}}

\item{resampling}{resampling method, see \code{gdalcubes::cubes_view}}

\item{maxcloud}{maximum cloud cover}
}
\value{
proxy data cube object
}
\description{
Create yearly composite images from STAC collection
}
\details{
If the \code{months} argument is provided, the lowest and highest number if used to define the compositing period.
}
