% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI_composite.R
\name{eo_VI_yearly.stac}
\alias{eo_VI_yearly.stac}
\title{Vegetation Index composites}
\usage{
eo_VI_yearly.stac(
  x,
  VI,
  endpoint,
  collection,
  stacOptions = list(limit = 1000, authOpt = list()),
  years,
  months,
  maxCloud = NULL,
  maskOptions = list(maskLyr = maskLyrName.stac(endpoint, collection), maskSnow = FALSE,
    maskWater = FALSE)
)
}
\arguments{
\item{x}{description}

\item{VI}{Vegetation Index (see details)}

\item{endpoint}{description}

\item{collection}{description#'}

\item{years}{numeric vector of years for which composite image must be created}

\item{months}{numeric vector of months for which composite image must be created for each years}

\item{maxCloud}{numeric, maximum cloud cover percentage}

\item{maskOptions}{description}
}
\value{
SpatRaster
}
\description{
Create composite image time series of a selected vegetation index
}
\details{
Limited functionalities for now
Implemented vegetation indices:
*NDVI = (nir-red)/(nir+red)
*EVI
*SAVI
*MSAVI
*NDMI
*NBR
*NBR2 = (swir1-swir2)/(swir1+swir2)
*NDSI = (green-swir1)/(green+swir1)
}
