% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_evaluation.R
\name{evaluate_matching}
\alias{evaluate_matching}
\title{Matching evaluation}
\usage{
evaluate_matching(match_object, covariate_overlap = TRUE, love_plot = TRUE)
}
\arguments{
\item{match_object}{description}

\item{covariate_overlap}{logical. Plot covariate density plots to evaluate covariate overlap.}

\item{love_plot}{logical. Plot Love plot.}
}
\description{
Methods for evaluating matching results
}
\details{
This function uses the [cobalt::bal.tab()] function to generate balance statistics from the [MatchIt::matchit()] call. 
If the argument `covariate_overlap` is set to `TRUE`, the function calls [cobalt::bal.plot()] and returns density plots for the all covariates for control and treatment groups.
If the argument `love_plot` is set to `TRUE`, a Love plot of Standardizes Mean Difference is generated with [cobalt::bal.plot()].
}
