% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatches.R
\name{extractFromRaster}
\alias{extractFromRaster}
\title{Extract from raster}
\usage{
extractFromRaster(r, xy, crs_xy)
}
\arguments{
\item{xy}{description}

\item{crs_xy}{description}

\item{name}{description}
}
\value{
data.table
}
\description{
Generic function to extract from SpatRaster or data cube object
}
