% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatches.R
\name{extractGeneric}
\alias{extractGeneric}
\title{Extract from generic raster}
\usage{
extractGeneric(
  x,
  y,
  fun = mean,
  method = "simple",
  reduce_time = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{SpatRaster or data cube}

\item{y}{SpatVector or sf. Additionally, if x is SpatRaster, see possibilities in \code{terra::extract}}

\item{fun}{function to summarize by line or polygon geometry}

\item{...}{Additional arguments passed to \code{terra::extract} or \code{gdalcubes::extract_geom}}
}
\value{
data.frame, matrix or SpatVector (see \code{terra::extract} or \code{gdalcubes::extract_geom})
}
\description{
Generic function to extract from SpatRaster or data cube object
}
