% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatches.R
\name{extractMatches}
\alias{extractMatches}
\title{Extract matches values from SpatRaster/data cube}
\usage{
extractMatches(r, obj, fun, method = "simple", ...)
}
\arguments{
\item{r}{a SpatRaster or data.cube raster object.}

\item{obj}{output of [matchCI()].}

\item{fun}{function to summarize extracted data by line or polygon geometries. See [terra::extract()].}

\item{method}{character. Method for extracting values for point geometries. See [terra::extract()].}

\item{...}{additional arguments to `fun`.}
}
\value{
data.table of extracted raster values for control-impact pairs.
}
\description{
Helper function to extract matches from SpatRaster or data.cube object
}
