% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcube_as_terra.R
\name{gdalcube_as_terra}
\alias{gdalcube_as_terra}
\title{Data cube to SpatRaster}
\usage{
gdalcube_as_terra(cube)
}
\arguments{
\item{cube}{data cube object}
}
\value{
a SpatRaster object
}
\description{
Convert a proxy data cube object to a terra SpatRaster object
}
\details{
Currently only works for data cubes with a single time dimension. To be improved.
}
