% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature.R
\name{get_feature}
\alias{get_feature}
\alias{get_feature,ANY,SpatRaster-method}
\alias{get_feature,ANY,SpatExtent-method}
\title{Extract STAC feature for specified spatial reference}
\usage{
get_feature(feature, x, assets, as_list = FALSE, add_datetime = TRUE, ...)

\S4method{get_feature}{ANY,SpatRaster}(feature, x, assets, as_list = FALSE, add_datetime = TRUE, ...)

\S4method{get_feature}{ANY,SpatExtent}(
  feature,
  x,
  assets,
  as_list = FALSE,
  add_datetime = TRUE,
  CRS,
  snap = "near",
  mask = FALSE,
  touches = TRUE,
  extend = FALSE,
  filename = "",
  ...
)
}
\arguments{
\item{feature}{description}

\item{x}{description}

\item{assets}{description}

\item{as_list}{description}

\item{add_datetime}{description}

\item{...}{description}
}
\value{
(list of) SpatRaster objects
}
\description{
Download STAC feature from collection and transform based on provided spatial reference
}
