% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature.R
\name{get_FeatureCollection}
\alias{get_FeatureCollection}
\title{Extract STAC FeatureCollection for specified spatial reference}
\usage{
get_FeatureCollection(FeatureCollection, x, assets, progress = FALSE, ...)
}
\arguments{
\item{x}{Spatial reference (see get_feature())}

\item{assets}{character vector of assets names to extract}

\item{progress}{should progress bar be shown}

\item{...}{additional arguments passed to get_feature}

\item{featureCollection}{STAC FeatureCollection (list)}
}
\value{
list of (list of) SpatRaster objects
}
\description{
Download STAC FeatureCollection from collection and transform based on provided spatial reference
}
