% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc.R
\name{lulc}
\alias{lulc}
\title{Extract Land USe / Land Cover data}
\usage{
lulc(
  x,
  year,
  endpoint = "https://planetarycomputer.microsoft.com/api/stac/v1",
  collection = "nasadem",
  assets = "elevation",
  signOpt = list(),
  ...
)
}
\arguments{
\item{x}{a SpatRaster or SpatExtent object, defining the area for which the LULC should be extracted}

\item{year}{description}

\item{endpoint}{description, defaults to Microsoft Planetary Computer}

\item{collection}{STAC collection, defaults to "io-lulc-9-class"}

\item{assets}{character defining STAC assets name, defaults to "data"}

\item{signOpt}{list with STAC authentication options passed to stac_auth (e.g. \code{list(key="abc")}), endpoint-specific}

\item{...}{additional parameters passed to \code{terra:project}}
}
\value{
SpatRaster object
}
\description{
Read LULC as SpatRaster
}
\details{
Designed to be flexible with regard to STAC endpoint and collection, but for now only tested for io-lulc-9-class on PlanetaryComputer 
io-lulc-9-class is only generated for the period 2017-2022, for years outside this range a warning will be returned and the neares year returned
}
