% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_format.R
\name{mapBand}
\alias{mapBand}
\title{Map band}
\usage{
mapBand(bands, collection, endpoint)
}
\arguments{
\item{bands}{description}

\item{collection}{description}

\item{endpoint}{description}
}
\value{
character vector
}
\description{
Map generic band name to collection/endpoint-specific band name
}
\details{
Limited number of collections/endpoints implemented, more to be added
}
