% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_composites.R
\name{maskLyrName.stac}
\alias{maskLyrName.stac}
\title{Mask layer name}
\usage{
maskLyrName.stac(endpoint, collection)
}
\arguments{
\item{endpoint}{STAC endpoint}

\item{collection}{STAC collection}
}
\value{
character or NULL
}
\description{
Function that defines metadata layer name from which mask can be derived for a collection at a STAC endpoint
}
\details{
Currently only implemented for collections "landsat-c2-l2" and "sentinel-2-l2a" at Planetary Computer. Returns NULL for all other values
}
