% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_matching.R
\name{matchCandidates}
\alias{matchCandidates}
\title{Matching candidates}
\usage{
matchCandidates(
  x,
  y,
  excludeBufferIn = 0,
  excludeBufferOut = 0,
  excludeOther = NULL
)
}
\arguments{
\item{x}{SpatVector: polygon of "impact" area}

\item{y}{SpatRaster: reference geometry}

\item{excludeBufferIn}{Numeric: Buffer (in m) inside polygon to be excluded}

\item{excludeBufferOut}{Numeric: Buffer (in m) outside polygon to be excluded}

\item{excludeOther}{SpatVector: other areas to be excluded}
}
\value{
SpatRaster object with candidate impact, control and excluded pixels
}
\description{
Identifies control-impact candidates for matching.
}
\details{
This function creates a raster object in which candidate impact pixels are labeled as 1,
candidate control pixels labeled as 0, and excluded pixels as NA.
}
