% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasadem.R
\name{nasadem}
\alias{nasadem}
\title{Extract Digital Elevation Model}
\usage{
nasadem(
  x,
  v = c("slope", "aspect"),
  unit = "degrees",
  aspectTransform = TRUE,
  endpoint = "https://planetarycomputer.microsoft.com/api/stac/v1",
  collection = "nasadem",
  assets = "elevation",
  authOpt = list(),
  ...
)
}
\arguments{
\item{x}{a SpatRaster or SpatExtent object, defining the area for which the DEM should be extracted}

\item{v}{description}

\item{unit}{description}

\item{aspectTransform}{description}

\item{endpoint}{description}

\item{collection}{STAC collection, defaults to "nasadem"}

\item{assets}{character defining STAC assets name, defaults to "elevation"}

\item{...}{additional parameters passed to ???}

\item{signOpt}{list with STAC authentication options passed to stac_auth (e.g. \code{list(key="abc")}), endpoint-specific}
}
\value{
SpatRaster object
}
\description{
Read NASA (SRTM) Digital Elevation Model as SpatRaster and calculate derivatives
}
