% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_distance.R
\name{osm_distance}
\alias{osm_distance}
\title{OpenStreetMap distance}
\usage{
osm_distance(x, what, ...)
}
\arguments{
\item{x}{SpatRaster or SpatVector for which distances are calculated}

\item{what}{Either "roads" or "places"}

\item{...}{Additional arguments to osm_distance_roads or osm_distance places}
}
\description{
Calculated shortest geographic distance to selected OpenStreetMap features
}
\details{
See osm_distance_roads, osm_distance places
}
