% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_distance.R
\name{osm_distance_places}
\alias{osm_distance_places}
\title{OpenStreetMap distance to places}
\usage{
osm_distance_places(
  x,
  values = "hamlet+",
  timeout = 25,
  name = NULL,
  osm_bbox = NULL,
  ...
)
}
\arguments{
\item{x}{SpatRaster or SpatVector for which distances are calculated}

\item{values}{Character vector of place categories to be included (see https://wiki.openstreetmap.org/wiki/Key:place) or a single place category with suffix "+" to include all categories including and above the selected category. Set to NULL to include all values.}

\item{timeout}{Numeric (seconds)}

\item{name}{Character of place name}

\item{osm_bbox}{Bounding box for OSM feature retrieval, as numeric vector or character string of administrative name (see \code{osmdata::getbb})}
}
\value{
SpatRaster or SpatVector
}
\description{
Calculated shortest geographic distance to OpenStreetMap place features
}
\details{
When a bounding box is not provided in the osm_bbox argument, the maximum required bounding box will be identified through an iterative search. This may be a more efficient option than providing an overly wide bounding box.
}
