% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_rast.R
\name{ref_rast}
\alias{ref_rast}
\alias{ref_rast,SpatVector,missing-method}
\alias{ref_rast,numeric,numeric-method}
\title{Create a reference SpatRaster object}
\usage{
ref_rast(x, y, crs, resolution, buffer = 0, round_coords = FALSE, ...)

\S4method{ref_rast}{SpatVector,missing}(x, y, crs, resolution, buffer = 0, round_coords = FALSE, ...)

\S4method{ref_rast}{numeric,numeric}(x, y, crs, resolution, buffer = 0, round_coords = FALSE, ...)
}
\arguments{
\item{x}{longitude (with western hemisphere longitudes negative) or a \code{SpatVector} object}

\item{y}{latitude (with southern hemisphere latitudes negative), or missing (if x is a \code{SpatVector} object)}

\item{crs}{description of the Coordinate Reference System in PROJ.4, WKT or authority:code notation.
Defaults to the UTM zone of center coordinates}

\item{buffer}{numeric vector of 1, 2, or 4 elements. Indicates by how much (in units of crs) the spatial extent of x/y should be enlarged on each side. 
A single positive integer when the distance in x and y to be added is equal; or 2 number (x, y), or four (lower x, upper x, lower y, upper y).}

\item{round_coords}{logical or integer. Should the coordinates of the reference SpatRaster be rounded.
If TRUE, coordinates are rounded to the nearest integer; if a positive numeric rounded to the corresponding decimal; 
if a negative integer rounded to the corresponding power of 10 (e.g., \code{round_corrs=2} round to the nearest 100).}

\item{...}{additional arguments passed on to \code{terra::rast}}
}
\value{
SpatRaster
}
\description{
Define a spatial reference spatRaster object based on a polygon of pair of coordinates, and additional spatial arguments.
}
\details{
Deprecated: replaced by create_control_candidates for creating spatial reference and identifying control/impact pixels
}
\examples{
ref_rast(30, 45, resolution=10, buffer=200, round_coords=-1)
}
