% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_format.R
\name{reqBands}
\alias{reqBands}
\title{Required bands for selected index}
\usage{
reqBands(index)
}
\arguments{
\item{index}{description}
}
\value{
character vector
}
\description{
Return (generic) band names required for calculating specified vegetation index
}
\details{
Possible values of \code{index} are "ndvi", "evi", "savi", "msavi", "ndmi", "nbr", "nbr2", "ndsi"
}
\seealso{
[cubeVI()]
}
