% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgrids.R
\name{SoilGrids2rast}
\alias{SoilGrids2rast}
\title{Extract most probable soil type from SoilGrids}
\usage{
SoilGrids2rast(product = "MostProbable")
}
\arguments{
\item{x}{reference SpatRaster}
}
\description{
Read most probable soil time from ISRIC SoilGrids data as SpatRaster
}
\details{
This is a quick draft that should be further developed based on the soilDB package
}
\references{
SoilGrids ISRIC CC-BY 4.0 License
}
