% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialBACI-package.R
\docType{package}
\name{spatialBACI-package}
\alias{spatialBACI}
\alias{spatialBACI-package}
\title{spatialBACI: Spatial Before-After-Control-Impact assessment}
\description{
Earth Observation-based to evaluate the impact of conservation or restoration actions following a (Before-After-)Control-Impact framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://space4restoration.github.io/spatialBACI/}
  \item \url{https://github.com/Space4Restoration/spatialBACI}
  \item Report bugs at \url{https://github.com/Space4Restoration/spatialBACI/issues}
}

}
\author{
\strong{Maintainer}: Jasper Van doninck \email{j.vandoninck@utwente.nl} (\href{https://orcid.org/0000-0003-2177-7882}{ORCID})

Other contributors:
\itemize{
  \item Louise Willemen (\href{https://orcid.org/0000-0003-1026-5865}{ORCID}) [contributor]
  \item Wietske Bijker (\href{https://orcid.org/0000-0002-4197-0295}{ORCID}) [contributor]
}

}
\keyword{internal}
