% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_auth.R
\name{stac_auth}
\alias{stac_auth}
\title{STAC authentication}
\usage{
stac_auth(x, endpoint, ...)
}
\arguments{
\item{x}{a STAC Feature or FeatureCollection}

\item{endpoint}{a STAC endpoint}

\item{...}{additional arguments specific to STAC endpoint}
}
\value{
description
}
\description{
Authenticate STAC items for specified STAC endpoint
}
\details{
Implemented endpoints: 
* Microsoft Planetary Computer
}
