% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multicollinearity.R
\name{test_multicollinearity}
\alias{test_multicollinearity}
\title{Test multicollinearity of matching covariates}
\usage{
test_multicollinearity(
  matching_input,
  colname.id,
  colname.treatment,
  colnames.ignore = NULL
)
}
\arguments{
\item{matching_input}{list with names elements "data" and "spat.ref". Output of [collate_matching_layers()].}

\item{colname.id}{character. Name of the column containing the unique ID of the spatial unit of analysis. Defaults to "cell" if `matching_input$spat.ref` is a SpatRaster.}

\item{colname.treatment}{character. Name of the column containing the treatment variable. Defaults to "treatment" if `matching_input$spat.ref` is a SpatRaster}

\item{colnames.ignore}{character vector. Name of columns (other than `colname.id` and `colname.treatment`) in `matching_input` to be ignored as matching covariates.}
}
\value{
list. Similar to `matching_input`, with optionally covariates in the data.table under "data" omitted.
}
\description{
The function provides a correlation plot of the (numeric) covariates, and iteratively reports (Generalized) Variance Inflation Factors of the different covariates.
After each iteration, the used is prompted whether a matching covariate should be omitted.
}
