% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI_composite.R
\name{VI_composite}
\alias{VI_composite}
\title{Vegetation Index composites}
\usage{
VI_composite(endpoint, collection, VI, spatRef, years, months, maxCloud)
}
\arguments{
\item{endpoint}{description}

\item{collection}{description}

\item{VI}{Vegetation Index (see details)}

\item{years}{numeric vector of years for which composite image must be created}

\item{months}{numeric vector of months for which composite image must be created for each years}

\item{maxCloud}{numeric, maximum cloud cover percentage}
}
\value{
SpatRaster
}
\description{
Create composite image time series of a selected vegetation index
}
\details{
Limited functionalities for now
Implemented vegetation indices:
*NDVI = (nir-red)/(nir+red)
*EVI
*SAVI
*MSAVI
*NDMI
*NBR
*NBR2 = (swir1-swir2)/(swir1+swir2)
*NDSI = (green-swir1)/(green+swir1)
}
