% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.bbox}
\alias{as.bbox}
\title{As bounding box}
\usage{
as.bbox(x, ...)
}
\arguments{
\item{x}{a SpatRaster, SpatVector or SpatExtent object}

\item{crs_to}{character. Description of the Coordinate Reference System of the created bounding box, in PROJ.4, WKT or authority:code notation. Defaults to longitude/latitude.}

\item{crs_from}{If x has no CRS attached (i.e., if x is a SpatExtend or numeric), a character description of the Coordinate Reference System corresponding to the calues of x. Defaults to longitude/latitude.}

\item{xy}{logical. If x is a numeric vector of length 4 (or 4 numeric values), set this parameter to TRUE if coordinates are in order (xmin, ymin, xmax, ymax) or FALSE if in order (xmin, xmax, ymin, ymax).}
}
\value{
numeric vector of length four
}
\description{
Create a Bounding Box from a SpatRaster, SpatVector, or SpatExtent object in target CRS. Bounding box is a 4-element vector in the order c(xmin, ymin, xmax, ymax)
}
