% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.bbox}
\alias{as.bbox}
\title{As bounding box}
\usage{
as.bbox(x, ...)
}
\arguments{
\item{x}{a SpatRaster, SpatVector or SpatExtent object}
}
\value{
numeric vector of length four
}
\description{
Create a Bounding Box from a SpatRaster or SpatExtent object in target CRS. Bounding box is a 4-element vector in the order c(xmin, ymin, xmax, ymax)
}
