% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.datetime}
\alias{as.datetime}
\title{As datetime}
\usage{
as.datetime(x, ...)
}
\arguments{
\item{x}{a numeric (or character?)}
}
\value{
a character string in STAC datetime format
}
\description{
Format to datetime
}
