% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.endpoint}
\alias{as.endpoint}
\title{As endpoint}
\usage{
as.endpoint(name)
}
\arguments{
\item{name}{character. Simplified endpoint name}
}
\value{
a character
}
\description{
Simplified representation of some STAC API endpoints
}
\details{
Defaults to "https://planetarycomputer.microsoft.com/api/stac/v1" (Planetary Computer). 

Possible values:
 * "lpdaac" = "https://cmr.earthdata.nasa.gov/stac/LPCLOUD",
 * "lpcloud" = "https://cmr.earthdata.nasa.gov/stac/LPCLOUD",
 * "planetarycomputer" = "https://planetarycomputer.microsoft.com/api/stac/v1",
 * "planetary computer" = "https://planetarycomputer.microsoft.com/api/stac/v1"
}
