% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_eval.R
\name{BACI_contrast}
\alias{BACI_contrast}
\title{Before-After-Control-Impact contrast}
\usage{
BACI_contrast(data, SpatRef = NULL)
}
\arguments{
\item{data}{data.table}

\item{SpatRef}{SpatRaster}
}
\value{
data.table
}
\description{
Calculate the Control-Impact contrast and p-value
}
\details{
After [del Rio] and [add reference]
}
