% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baviaanskloof-data.R
\docType{data}
\name{baviaanskloof}
\alias{baviaanskloof}
\title{Baviaanskloof revegetation data}
\format{
A SpatVector object
}
\usage{
data(baviaanskloof)
}
\description{
Data on large scale planting in Baviaanskloof, South Africa.
}
\details{
Data used and described in detail in del Río-Mena et al.
}
\examples{
data(baviaanskloof)
baviaanskloof <- unwrap(baviaanskloof)
plot(baviaanskloof, "Planting_date")

}
\references{
del Río-Mena, T., Willemen, L., Vrieling, A., Snoeys, A., Nelson, A., 2021. Long-term assessment of ecosystem services at ecological restoration sites using Landsat time series. PLOS ONE 16, e0243020. <https://doi.org/10.1371/journal.pone.0243020>
}
\keyword{datasets}
