% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_mask.R
\name{bitmask_values}
\alias{bitmask_values}
\title{Landsat L2 QA bitmap values}
\usage{
bitmask_values(
  nBit = 8,
  fill = TRUE,
  dilCloud = TRUE,
  cirrus = TRUE,
  cloud = TRUE,
  cloudShadow = TRUE,
  snow = FALSE,
  water = FALSE
)
}
\arguments{
\item{fill}{description}

\item{dilCloud}{description}

\item{cirrus}{description}

\item{cloud}{description}

\item{cloudShadow}{description}

\item{snow}{description}

\item{water}{description}

\item{nbit}{description}
}
\value{
integer vector
}
\description{
Define the values of values to be masked from Landsat L2 products based on specified QA bits
}
