% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_metrics.R
\name{calc_ts_metrics.cube}
\alias{calc_ts_metrics.cube}
\title{Calculate time series metrics from a data cube}
\usage{
calc_ts_metrics.cube(cube, average = 1, trend = 1)
}
\arguments{
\item{cube}{data cube with a t dimension.}

\item{average}{numeric. If not 0, calculate the time series average}

\item{trend}{numeric. If not 0, calculate the time series trend, and the intercept at tmin and tmax}
}
\value{
data cube
}
\description{
Calculate average and trend from a multitemporal data cube
}
