% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_matching.R
\name{collate_matching_layers}
\alias{collate_matching_layers}
\title{Collate matching layers}
\usage{
collate_matching_layers(
  x,
  vars_list,
  colname.id = "ID",
  colnames.ignore = NULL
)
}
\arguments{
\item{x}{SpatRaster or SpatVector}

\item{vars_list}{List if matching variables. See Details.}

\item{colname.id}{if x is a SpatVector, the column name identifying the ID allowing merging of different datasets}

\item{colnames.ignore}{if x is a SpatVector, vector of column names in vars_list to ignore}
}
\value{
description
}
\description{
Process different matching layers in compatible format
}
\details{
If x is a SpatRaster, vars_list should have as list elements SpatRaster, data cube objects. 
If x is a SpatVector, vars_list should have as list elements SpatVector, data.frame, or data.table objects.
}
