% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_matching.R
\name{collate_matching_layers}
\alias{collate_matching_layers}
\title{Collate matching layers}
\usage{
collate_matching_layers(x, ...)
}
\arguments{
\item{x}{spatRaster providing spatial reference}

\item{...}{Matching layers. See Details.}
}
\description{
Process different matching layers in compatible format
}
\details{
All matching layers should be provided as additional arguments (\code{...}). Matching layers can be provided as spatRaster, data cube, or a character (which will be interpreted as a readable file name).
}
