% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_cube.R
\name{eo_cube.stac}
\alias{eo_cube.stac}
\title{EO cube from STAC}
\usage{
eo_cube.stac(
  x,
  endpoint,
  collection,
  stacOptions = list(),
  t0,
  t1,
  dt = "P1D",
  asset_names = NULL,
  joint_assets = TRUE,
  maxCloud = NULL,
  maskOptions = list(NULL),
  aggregation = "first",
  resampling = "bilinear",
  harm_bandnames = TRUE
)
}
\arguments{
\item{x}{spatRaster. Defines spatial reference of the output data cube}

\item{endpoint}{character. STAC endpoint}

\item{collection}{character. STAC collection}

\item{stacOptions}{list. Optional arguments to be passed to eo_stac_search, e.g. authentication options}

\item{t0}{start date as numeric, character or Date}

\item{t1}{end date as numeric, character or Date}

\item{dt}{size of pixels in time-direction, expressed as ISO8601 period string}

\item{asset_names}{character vector of asset (band) names to be used, or NULL for all asset names with "eo:bands" attributes}

\item{joint_assets}{logical indicating if only the assets (bands) present in all features should be used, if argument asset_names is NULL}

\item{maxCloud}{maximum cloud cover [0-100] of images to be included in data cube, only relevant for collections with "eo:cloud_cover" attributes}

\item{maskOptions}{list specifying layer used to create mask and masking options (see eo_mask.cube)}

\item{resampling}{see gdalcubes::cube_view}

\item{harm_bandnames}{logical indicating whether endpoint/collection-specific band names should be changed to generic band names}

\item{aggregations}{see gdalcubes::cube_view}
}
\value{
proxy data cube object
}
\description{
Create a data cube for EO images retrieved from a STAC catalog.
}
\details{
t0 and t1 can be provided as numeric in format yyyymmdd, as a character in the format "yyyymmdd" or "yyyy-mm-dd", or as a Date object. 
For now, only date is implemented, not time.
}
