% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_stac_search.R
\name{eo_stac_search.PlanetaryComputer}
\alias{eo_stac_search.PlanetaryComputer}
\title{Search STAC EO data for Planetary Computer}
\usage{
eo_stac_search.PlanetaryComputer(
  collection = NULL,
  ids = NULL,
  bbox = NULL,
  datetime = NULL,
  intersects = NULL,
  limit = NULL,
  eocc = NULL,
  sat.orbit.state = NULL,
  sat.orbit = NULL,
  authOpt = list()
)
}
\arguments{
\item{collection}{STAC collection}

\item{ids}{description}

\item{bbox}{description}

\item{datetime}{description}

\item{intersects}{description}

\item{limit}{description}

\item{eocc}{maximum cloud cover (for "eo:cloud_cover" field)}

\item{authOpt}{list specifying STAC authentication options (e.g. list(key="abc"))}
}
\value{
STACItemCollection
}
\description{
Search STAC EO data for Planetary Computer
}
