% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatches.R
\name{extractMatches}
\alias{extractMatches}
\title{Extract data for matches}
\usage{
extractMatches(matches, effect, before, after)
}
\arguments{
\item{matches}{output of \code{matchCI}}

\item{effect}{SpatRaster of effect values}

\item{before}{SpatRaster of before values}

\item{after}{SpatRaster of after values}
}
\description{
Extract the "before" and "after", or "effect", values for the matching pairs
}
\seealso{
[matchCI()]
}
