% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatches.R
\name{extractMatches}
\alias{extractMatches}
\title{Extract data from raster for matches}
\usage{
extractMatches(matches, effect, before, after, band = NULL)
}
\arguments{
\item{matches}{output of \code{matchCI}}

\item{effect}{SpatRaster of effect values}

\item{before}{SpatRaster of before values}

\item{after}{SpatRaster of after values}

\item{band}{character. If before/after contain more than one band, this parameter can be used to select the band to be extracted}
}
\description{
Extract the "before" and "after", or "effect", values from a SpatRaster or data cube for the provided matching pairs
}
