% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_matching.R
\name{matchCI}
\alias{matchCI}
\title{Control-impact matching}
\usage{
matchCI(
  cands,
  matchVars,
  colname.id = "ID",
  colname.treatment = "treatment",
  ratio = 1,
  replace = FALSE,
  method = "nearest",
  distance = "glm",
  link = "logit",
  eval = FALSE,
  asmd_warn = NULL,
  asmd_error = NULL,
  ...
)
}
\arguments{
\item{cands}{spatRaster or spatVector object. See Details.}

\item{matchVars}{list of matching variables. See Details.}

\item{ratio}{control/impact ratio. See \code{matchit} function}

\item{replace}{See \code{matchit} function}

\item{method}{See \code{matchit} function}

\item{distance}{See \code{matchit} function}

\item{link}{See \code{matchit} function}

\item{eval}{logical. Should matching be evaluated? Function will return NULL if matching is rejected}

\item{asmd_warn}{numeric. Function will output a warning message if Absolute Standardized Mean Difference is above this value}

\item{asmd_error}{numeric. Function will output an error if Absolute Standardized Mean Difference is above this value}

\item{...}{additional inputs to \code{matchit} function}
}
\value{
data table with matched control-impact units. Column names are "ID", "treatment", and "subclass".
}
\description{
Match control units to impact units
}
\details{
This function uses the provided matching layers to match the provided candidate and impact units 
using the MatchIt package and the matching options provided therein.

If cands is a spatRaster, impact pixels should be labelled 1, candidate control pixels labelled 0, and pixels to be excluded labelled NA.
In that case, matching variables in \code{matchvars} are to be provided as a single (multilayer) terra spatRaster object, gdalcubes data cube object, character pathname (which will be assumed to refer to a raster imagee), or as a list of these classes.
If cands is a spatVector, attributes should include "ID", "treatment" (1 for impact sites and 0 for candidate control sites). All remaining columns will be interpreted as matching variables
}
