% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_matching.R
\name{matchCI}
\alias{matchCI}
\title{Control-impact matching}
\usage{
matchCI(
  cands,
  matchlyrs,
  eval = FALSE,
  cols = c("subclass", "x", "y", "treatment"),
  ...
)
}
\arguments{
\item{cands}{Control-Impact candidates - SpatRaster - 0=control, 1=impact, NA=exclude}

\item{matchlyrs}{multilayer SpatRaster (same geometry as CI_cand) of matching variables}

\item{eval}{should matching be evaluated, return NULL if matching is rejected}

\item{cols}{columns to be returned in the data table}

\item{...}{additional inputs to \code{matchit} function}
}
\value{
data table with matched control-impact units
}
\description{
Match control units to impact units
}
\details{
This function uses the provided matching layers to match the provided candidate and impact units 
using the MatchIt package and the matching options provided therein.
}
