% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgrids.R
\name{SoilGrids2rast}
\alias{SoilGrids2rast}
\title{Read SoilGrids layer as spatRaster}
\usage{
SoilGrids2rast(product = "MostProbable")
}
\arguments{
\item{product}{SoilGrids layer to be read as spatRaster, defaults to "MostProbable"}
}
\description{
Read SoilGrids layer as spatRaster
}
\references{
SoilGrids ISRIC CC-BY 4.0 License
}
